Corrigé quelques données des tests WIT suite au changement de référentiel des rues (IFOII -> FiDoR/RefInf)
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@118475 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/web-it/src/test/java/ch/vd/uniregctb/webservice/party4/PartyWebServiceTest.java
@@ -267,7 +267,7 @@ public class PartyWebServiceTest extends AbstractPartyWebServiceTest {
 
 		final AddressInformation courrierInfo2 = courrier2.getAddressInformation();
 		assertNotNull(courrierInfo2);
-		assertEquals(new Integer(141554), courrierInfo2.getStreetId());
+		assertEquals(new Integer(1134061), courrierInfo2.getStreetId());
 		assertEquals("12", courrierInfo2.getHouseNumber());
 		assertEquals(Integer.valueOf(1000), courrierInfo2.getSwissZipCodeId());
 		assertEquals(Long.valueOf(1753), courrierInfo2.getSwissZipCode());
